requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
isort==5.10.1
kombu==5.2.4
mccabe==0.6.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
